% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_network.R
\name{assign_genes_to_states}
\alias{assign_genes_to_states}
\title{Assigns genes to states}
\usage{
assign_genes_to_states(
  matrix,
  dynamic_object,
  method = "active_expression",
  p_value = 0.05,
  pThresh_DE = 0.05,
  active_thresh = 0.33,
  toScale = FALSE,
  forceGenes = TRUE
)
}
\arguments{
\item{matrix}{genes-by-cells expression matrix}

\item{dynamic_object}{individual path result of running define_states}

\item{method}{method of assigning state genes, either "active_expression" (looks for active expression in state) or "DE" (looks for differentially expressed genes per state)}

\item{p_value}{pval threshold if gene is dynamically expressed}

\item{pThresh_DE}{pval if gene is differentially expressed. Ignored if method is active_expression.}

\item{active_thresh}{value between 0 and 1. Percent threshold to define activity}

\item{toScale}{whether or not to scale the data}

\item{forceGenes}{whether or not to rescue orphan dyanmic genes, forcing assignment into state with max expression.}
}
\value{
states a list detailing genes active in each state
}
\description{
Assigns genes to states
}
