% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_network.R
\name{assign_genes_to_states_simple}
\alias{assign_genes_to_states_simple}
\title{Assigns genes to states just based on which mean is maximal}
\usage{
assign_genes_to_states_simple(
  matrix,
  dynamic_object,
  num_states = 3,
  pThresh = 0.01,
  toScale = FALSE,
  key_word = "state_"
)
}
\arguments{
\item{matrix}{expression matrix}

\item{dynamic_object}{result of running findDynGenes}

\item{num_states}{num_states}

\item{pThresh}{pThresh}

\item{toScale}{toScale}

\item{key_word}{key_word}
}
\value{
data.frame of dynamically expressed genes, cluster, peakTime, ordered by peaktime
}
\description{
Assigns genes to states just based on which mean is maximal
}
