% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gene_rank.R
\name{calculate_gene_rank}
\alias{calculate_gene_rank}
\alias{calculate_gene_rank,Network-method}
\alias{calculate_gene_rank,Seurat-method}
\alias{calculate_gene_rank,CSNObject-method}
\alias{calculate_gene_rank,data.frame-method}
\title{Calculate gene rank}
\usage{
calculate_gene_rank(object, ...)

\S4method{calculate_gene_rank}{Network}(
  object,
  regulators = NULL,
  targets = NULL,
  directed = FALSE,
  method = c("page_rank", "degree_distribution"),
  ...
)

\S4method{calculate_gene_rank}{Seurat}(
  object,
  network = NULL,
  celltypes = NULL,
  regulators = NULL,
  targets = NULL,
  directed = FALSE,
  method = c("page_rank", "degree_distribution"),
  ...
)

\S4method{calculate_gene_rank}{CSNObject}(object, ...)

\S4method{calculate_gene_rank}{data.frame}(
  object,
  regulators = NULL,
  targets = NULL,
  directed = FALSE,
  method = c("page_rank", "degree_distribution"),
  ...
)
}
\arguments{
\item{object}{Network object}

\item{...}{Additional arguments}

\item{regulators}{Character vector, regulators to include}

\item{targets}{Character vector, targets to include}

\item{directed}{Logical, whether the network is directed}

\item{method}{Character, ranking method: "page_rank" or "degree_distribution"}

\item{network}{Name of the network to use.}

\item{celltypes}{Character vector specifying which states/cell types to include; NULL for all.}
}
\value{
Data frame with gene ranks

The Network object with gene ranks stored in \code{params$gene_rank}.

The Seurat object with gene ranks stored in each Network's \code{params$gene_rank}.
}
\description{
Calculate gene rank
}
