% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethod.R
\name{coef.CSNObject}
\alias{coef.CSNObject}
\title{Get fitted coefficients}
\usage{
\method{coef}{CSNObject}(object, network = DefaultNetwork(object), celltypes = NULL, ...)
}
\arguments{
\item{object}{CSNObject object}

\item{network}{Name of the network to use.}

\item{celltypes}{Celltypes to plot.}

\item{...}{Other parameters}
}
\value{
Return the fitted coefficients
}
\description{
Get fitted coefficients
}
