% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_analysis.R
\name{compare_celltype_rewiring}
\alias{compare_celltype_rewiring}
\title{Compare rewiring between static cell type networks}
\usage{
compare_celltype_rewiring(
  object,
  network = DefaultNetwork(object),
  celltypes = NULL,
  weight_column = "weight"
)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{network}{Static network name. Defaults to \code{DefaultNetwork(object)}.}

\item{celltypes}{Optional subset of cell types.}

\item{weight_column}{Weight column used when exporting networks.}
}
\value{
A list with pair summaries and per-TF rewiring tables.
}
\description{
Compare rewiring between static cell type networks
}
