% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{compile_states}
\alias{compile_states}
\title{compile_states}
\usage{
compile_states(networks_list)
}
\arguments{
\item{networks_list}{list of results of running assign_network on each path.}
}
\value{
a compiled list of network genes
}
\description{
compile networks based on matching network label in result of running assign_network.
}
