% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_doc.R
\name{CSNObject-methods}
\alias{CSNObject-methods}
\alias{$,CSNObject-method}
\alias{DefaultNetwork,CSNObject-method}
\alias{DefaultNetwork<-,CSNObject-method}
\alias{EdgeUniqueness,CSNObject-method}
\alias{export_csn,CSNObject-method}
\alias{ExtractGenesTransition,CSNObject-method}
\alias{GeneRanks,CSNObject-method}
\alias{get_network_graph,CSNObject-method}
\alias{get_tf_network,CSNObject-method}
\alias{GetNetwork,CSNObject-method}
\alias{metrics,CSNObject-method}
\alias{NetworkGraph,CSNObject-method}
\alias{NetworkModules,CSNObject-method}
\alias{NetworkParams,CSNObject-method}
\alias{NetworkRegions,CSNObject-method}
\alias{NetworkTFs,CSNObject-method}
\alias{Params,CSNObject-method}
\alias{plot_gof,CSNObject-method}
\alias{plot_module_metrics,CSNObject-method}
\alias{plot_network_graph,CSNObject-method}
\alias{plot_tf_network,CSNObject-method}
\alias{ShortestPaths,CSNObject-method}
\title{Methods for CSNObject objects}
\description{
S4 methods defined for \code{\link[=CSNObject-class]{CSNObject}} objects.
}
\keyword{internal}
