% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state.R
\name{CSNObject}
\alias{CSNObject}
\title{Show a compact summary of CSNObject state stored in a Seurat object}
\usage{
CSNObject(object, ...)
}
\arguments{
\item{object}{A \code{Seurat} object initialized by \code{multiCSN}.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Show a compact summary of CSNObject state stored in a Seurat object
}
