% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{DefaultNetwork}
\alias{DefaultNetwork}
\alias{DefaultNetwork<-}
\alias{DefaultNetwork,Seurat-method}
\alias{DefaultNetwork<-,Seurat-method}
\title{Get active network}
\usage{
DefaultNetwork(object, ...)

DefaultNetwork(object, ...)

DefaultNetwork(object, ...) <- value

\S4method{DefaultNetwork}{Seurat}(object, ...)

\S4method{DefaultNetwork}{Seurat}(object) <- value
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}

\item{value}{Name of the network to activate.}
}
\description{
Get active network

Get active network name
}
