% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_intersection.R
\name{dynamic_windowing}
\alias{dynamic_windowing}
\title{Legacy dynamic windowing wrapper}
\usage{
dynamic_windowing(
  meta_data,
  pseudotime_column = "pseudotime",
  group_column = "cluster"
)
}
\arguments{
\item{meta_data}{Metadata data.frame.}

\item{pseudotime_column}{Pseudotime column.}

\item{group_column}{Group column.}
}
\value{
Named list of state cells.
}
\description{
Legacy dynamic windowing wrapper
}
