% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_comparison.R
\name{edge_uniqueness}
\alias{edge_uniqueness}
\alias{edge_uniqueness,list-method}
\alias{edge_uniqueness,Seurat-method}
\alias{edge_uniqueness,CSNObject-method}
\title{Edge uniqueness across multiple GRNs}
\usage{
edge_uniqueness(object, ...)

\S4method{edge_uniqueness}{list}(object, tfs, weight_column)

\S4method{edge_uniqueness}{Seurat}(
  object,
  network = NULL,
  celltypes = NULL,
  tfs = NULL,
  weight_column = "weight",
  weight_cutoff = NULL
)

\S4method{edge_uniqueness}{CSNObject}(object, ...)
}
\arguments{
\item{object}{Named list of GRN data.frames (regulator, target, weight) or CSNObject.}

\item{...}{Passed to methods.}

\item{tfs}{If \code{NULL}, uses \code{object@metadata$tfs}.}

\item{weight_column}{Column name for edge weights.}

\item{network}{Name of dynamic network in \code{object@networks}.}

\item{celltypes}{States to include; \code{NULL} for all.}

\item{weight_cutoff}{Passed to \code{export_csn}.}
}
\description{
Edge uniqueness across multiple GRNs
}
