% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{EdgeUniqueness}
\alias{EdgeUniqueness}
\alias{EdgeUniqueness,Seurat-method}
\title{Get stored edge_uniqueness results from CSNObject}
\usage{
EdgeUniqueness(object, ...)

\S4method{EdgeUniqueness}{Seurat}(object, network = DefaultNetwork(object), ...)
}
\arguments{
\item{object}{The input data.}

\item{...}{Arguments for other methods}

\item{network}{Name of the network.}
}
\description{
Get stored edge_uniqueness results from CSNObject
}
