% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/network_perturbation.R
\name{embedPerturbation}
\alias{embedPerturbation}
\alias{embedPerturbation,Seurat-method}
\alias{embedPerturbation,Network-method}
\title{Embed perturbation}
\usage{
embedPerturbation(
  object,
  reduction_method = "umap",
  dims = 2,
  scale = TRUE,
  cores = 1,
  seed = 1
)

embedPerturbation(
  object,
  reduction_method = "umap",
  dims = 2,
  scale = TRUE,
  cores = 1,
  seed = 1
)

\S4method{embedPerturbation}{Seurat}(
  object,
  reduction_method = "umap",
  dims = 2,
  scale = TRUE,
  cores = 1,
  seed = 1
)

\S4method{embedPerturbation}{Network}(
  object,
  reduction_method = "umap",
  dims = 2,
  scale = TRUE,
  cores = 1,
  seed = 1
)
}
\arguments{
\item{object}{A Network or CSNObject object}

\item{reduction_method}{Character, dimensionality reduction method ("umap", "tsne", "pca")}

\item{dims}{Number of dimensions to compute}

\item{scale}{Whether to scale the data before reduction}

\item{cores}{Number of cores for parallel processing}

\item{seed}{Random seed for reproducibility}
}
\value{
Updated object with embedded perturbation results
}
\description{
Embed perturbation

Embed perturbation results
}
