% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{ExtractGenesTransition}
\alias{ExtractGenesTransition}
\alias{ExtractGenesTransition,Seurat-method}
\title{Get stored extract_genes_transition results from CSNObject}
\usage{
ExtractGenesTransition(object, ...)

\S4method{ExtractGenesTransition}{Seurat}(object, network = DefaultNetwork(object), ...)
}
\arguments{
\item{object}{The input data.}

\item{...}{Arguments for other methods}

\item{network}{Name of the stored network whose transition results are returned;
defaults to \code{DefaultNetwork(object)}.}
}
\description{
Get stored extract_genes_transition results from CSNObject
}
