% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motifs.R
\name{find_motifs}
\alias{find_motifs}
\alias{find_motifs,Seurat-method}
\alias{find_motifs,CSNObject-method}
\title{Scan for motifs in candidate regions}
\usage{
find_motifs(object, ...)

\S4method{find_motifs}{Seurat}(
  object,
  pfm,
  genome,
  motif_tfs = NULL,
  backend = c("signac", "motifmatchr"),
  verbose = TRUE,
  ...
)

\S4method{find_motifs}{CSNObject}(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}

\item{pfm}{A PFMatrixList object with position weight matrices.}

\item{genome}{A BSgenome object.}

\item{motif_tfs}{A data frame matching motifs with TFs. First column motif, second TF.}

\item{backend}{Motif scanning backend. `"signac"` uses `Signac::AddMotifs()`;
`"motifmatchr"` uses `motifmatchr::matchMotifs()` directly and wraps the
result into a Signac `Motif` object.}

\item{verbose}{Display messages.}
}
\value{
A CSNObject object with updated motif info.
}
\description{
Scan for motifs in candidate regions
}
\examples{
\dontrun{
# Supply a Seurat object containing RNA and chromatin accessibility assays.
data(motifs)

object <- Seurat::NormalizeData(object, assay = "RNA", verbose = FALSE)
object <- Signac::RunTFIDF(object, assay = "peaks", verbose = FALSE)
object <- initiate_object(
  object,
  group.by = "CellType",
  rna_assay = "RNA",
  peak_assay = "peaks",
  verbose = FALSE
)

genome <- getExportedValue("BSgenome.Hsapiens.UCSC.hg38", "BSgenome.Hsapiens.UCSC.hg38")
object <- find_motifs(
  object,
  pfm = motifs,
  genome = genome,
  backend = "motifmatchr",
  verbose = FALSE
)
}
}
