% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks_near_genes.R
\name{find_peaks_near_genes}
\alias{find_peaks_near_genes}
\title{Find peaks or regions near gene body or TSS}
\usage{
find_peaks_near_genes(
  peaks,
  genes,
  sep = c("-", "-"),
  method = c("Signac", "GREAT"),
  upstream = 1e+05,
  downstream = 0,
  extend = 1e+06,
  only_tss = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{peaks}{A \code{GRanges} object with peak regions.}

\item{genes}{A \code{GRanges} object with gene coordinates.}

\item{sep}{Vector of separators to use for genomic string.}

\item{method}{Character specifying the method to link peak overlapping motif regions to nearby genes.
One of 'Signac' or 'GREAT'.}

\item{upstream}{Integer defining the distance upstream of the gene/TSS to consider.}

\item{downstream}{Integer defining the distance downstream of the gene/TSS to consider.}

\item{extend}{Integer defining the distance from the upstream and downstream of the
basal regulatory region. Used only when method is 'GREAT'.}

\item{only_tss}{Logical value. Measure distance from the TSS (\code{TRUE})
or from the entire gene body (\code{FALSE}).}

\item{verbose}{Logical value. Display messages}
}
\value{
A sparse binary Matrix with gene/peak matches.
}
\description{
Find peaks or regions near gene body or TSS
}
