% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_srm2}
\alias{fit_srm2}
\title{Fit a sparse regression model}
\usage{
fit_srm2(
  formula,
  data,
  verbose = TRUE,
  max_support_size = NULL,
  min_improvement = 1e-10
)
}
\arguments{
\item{formula}{An object of class `formula` describing the model.}

\item{data}{A data frame containing the variables in the model.}

\item{verbose}{Whether to show warning messages.}

\item{max_support_size, min_improvement}{See [inferCSN::fit_greedy_l0()].}
}
\value{
A list containing model, fit metrics, and nonzero coefficients.
}
\description{
Fits a sparse regression model using [inferCSN::fit_greedy_l0()].
}
\examples{
data("example_matrix", package = "inferCSN")
df <- as.data.frame(example_matrix)
fit_srm2(g1 ~ ., data = df)
}
