% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{GeneRanks}
\alias{GeneRanks}
\alias{GeneRanks,Seurat-method}
\alias{GeneRanks,Network-method}
\title{Get gene ranks from Network/CSNObject}
\usage{
GeneRanks(object, ...)

\S4method{GeneRanks}{Seurat}(object, network = DefaultNetwork(object), celltypes = NULL, ...)

\S4method{GeneRanks}{Network}(object, ...)
}
\arguments{
\item{object}{The input data.}

\item{...}{Arguments for other methods}

\item{network}{Name of the network.}

\item{celltypes}{Cell types to get ranks for; NULL for all.}
}
\description{
Get gene ranks from Network/CSNObject
}
