% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/get_attribute.R
\name{get_attribute}
\alias{get_attribute}
\alias{get_attribute,Seurat-method}
\alias{get_attribute,CSNObject-method}
\title{Get any attribute from a CSNObject object}
\usage{
get_attribute(object, ...)

get_attribute(object, ...)

\S4method{get_attribute}{Seurat}(
  object,
  celltypes = NULL,
  active_network = NULL,
  attribute = c("genes", "tfs", "peaks", "regulators", "targets", "celltypes", "cells",
    "modules", "coefficients"),
  ...
)

\S4method{get_attribute}{CSNObject}(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}

\item{celltypes}{A character vector specifying the celltypes to get attributes for.}

\item{active_network}{A character string specifying the active network.}

\item{attribute}{Attribute to get: genes, tfs, peaks, regulators, targets, celltypes, cells, modules, coefficients.}
}
\value{
A character vector of regulatory genes

A character vector or list of regulatory genes
}
\description{
Get any attribute from a CSNObject object

Get any attribute from a CSNObject
}
