% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_adapters.R
\name{get_layer_data}
\alias{get_layer_data}
\title{Read assay layer data}
\usage{
get_layer_data(object, ...)
}
\arguments{
\item{object}{A Seurat object or a legacy CSNObject.}

\item{...}{Arguments passed to SeuratObject::LayerData.}
}
\value{
An assay layer matrix.
}
\description{
Read assay layer data
}
