% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/generics.R, R/utils_plots.R
\name{get_network_graph}
\alias{get_network_graph}
\alias{get_network_graph,Seurat-method}
\title{Get network graph}
\usage{
get_network_graph(object, ...)

get_network_graph(object, ...)

\S4method{get_network_graph}{Seurat}(
  object,
  network = NULL,
  celltypes = NULL,
  graph_name = "module_graph",
  rna_assay = "RNA",
  rna_layer = "data",
  umap_method = c("weighted", "coef", "corr", "none"),
  features = NULL,
  seed = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{CSNObject or Network}

\item{...}{Additional arguments for \code{\link[uwot]{umap}}.}

\item{network}{Name of the network to use.}

\item{celltypes}{Celltypes to plot.
If \code{NULL}, all celltypes are plotted.}

\item{graph_name}{Name of the graph.}

\item{rna_assay}{Name of the RNA assay.}

\item{rna_layer}{Name of the RNA slot to use.}

\item{umap_method}{Method to compute edge weights for UMAP:
* \code{'weighted'} - Correlation weighted by GRN coefficient.
* \code{'corr'} - Only correlation.
* \code{'coef'} - Only GRN coefficient.
* \code{'none'} - Don't compute UMAP and create the graph directly
from modules.}

\item{features}{Features to use to create the graph. If \code{NULL}
uses all features in the network.}

\item{seed}{Random seed for UMAP computation}

\item{verbose}{Print messages.}
}
\value{
A CSNObject object.
}
\description{
Get network graph

Get network graph

Compute network graph embedding using UMAP.
}
