% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_adapters.R
\name{get_peak_annotation}
\alias{get_peak_annotation}
\title{Read peak assay annotations}
\usage{
get_peak_annotation(object, peak_assay)
}
\arguments{
\item{object}{A Seurat object or a legacy CSNObject.}

\item{peak_assay}{Name of the peak assay.}
}
\value{
Genomic annotations, or NULL if unavailable.
}
\description{
Read peak assay annotations
}
