% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pseudotime.R
\name{get_pseudotime}
\alias{get_pseudotime}
\alias{get_pseudotime.default}
\alias{get_pseudotime.Seurat}
\title{Get pseudotime information}
\usage{
get_pseudotime(object, ...)

\method{get_pseudotime}{default}(
  object,
  meta_data = NULL,
  embeddings = NULL,
  cluster_column = "cluster",
  cores = 1,
  seed = 1,
  start_cluster = NULL,
  end_cluster = NULL,
  verbose = TRUE,
  ...
)

\method{get_pseudotime}{Seurat}(
  object,
  assay = "RNA",
  layer = "data",
  cluster_column = "cluster",
  reduction = "umap",
  start_cluster = NULL,
  end_cluster = NULL,
  ...
)
}
\arguments{
\item{object}{The input data (matrix, Seurat, etc.)}

\item{...}{Arguments for other methods}

\item{meta_data}{Input meta data.}

\item{embeddings}{Embeddings information}

\item{cluster_column}{The column used for slingshot.}

\item{cores}{Number of cores to use for parallel processing. Default is 1.}

\item{seed}{Random seed for reproducibility. Default is 1.}

\item{start_cluster}{The start cluster.}

\item{end_cluster}{The end cluster.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{assay}{The assay used for slingshot.}

\item{layer}{The layer used for slingshot.}

\item{reduction}{The reduction used for slingshot.}
}
\value{
Pseudotime information corresponding to cells
}
\description{
Get pseudotime information
}
