% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/generics.R, R/utils_plots.R
\name{get_tf_network}
\alias{get_tf_network}
\alias{get_tf_network,Seurat-method}
\title{Get TF network}
\usage{
get_tf_network(object, ...)

get_tf_network(object, ...)

\S4method{get_tf_network}{Seurat}(
  object,
  celltypes = NULL,
  tfs = NULL,
  features = NULL,
  network = NULL,
  graph = "module_graph",
  order = 3,
  keep_all_edges = FALSE,
  verbose = TRUE,
  cores = 1,
  ...
)
}
\arguments{
\item{object}{CSNObject or Network}

\item{...}{Additional arguments}

\item{celltypes}{Celltypes to plot.
If \code{NULL}, all celltypes are plotted.}

\item{tfs}{The transcription factors to center around.}

\item{features}{Features to use. If \code{NULL} uses all features in the graph.}

\item{network}{Name of the network to use.}

\item{graph}{Name of the graph.}

\item{order}{Integer indicating the maximal order of the graph.}

\item{keep_all_edges}{Logical, whether to maintain all edges to each leaf
or prune to the strongest overall connection.}

\item{verbose}{Logical. Whether to print messages.}

\item{cores}{Logical. Whether to parallelize the computation with \code{\link[foreach]{foreach}}.}
}
\value{
A CSNObject object.
}
\description{
Get TF network

Get TF network

Get sub-network centered around one TF.
}
