% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{heatmap_by_treatment_group}
\alias{heatmap_by_treatment_group}
\title{Useful plotting function to plot heatmap of module expression across time with pre-split matrix}
\usage{
heatmap_by_treatment_group(
  expList,
  meta_data,
  pseudotime_column = "pseudotime",
  toScale = T,
  limits = c(0, 5),
  smooth = TRUE,
  order_by = "WAG",
  thresh_on = 0.02,
  fontSize = 8,
  anno_colors = NA
)
}
\arguments{
\item{expList}{list of expression matrices}

\item{meta_data}{sample table}

\item{pseudotime_column}{column in sample table containing pseudotime annotation}

\item{toScale}{whether or not to scale expression}

\item{limits}{limits on expression}

\item{smooth}{whether or not to smooth expression across pseudotime for cleaner plotting}

\item{order_by}{name in expList that is used to order rows in the heatmap}

\item{thresh_on}{threshold expression is considered on, used in ordering the rows}

\item{fontSize}{heatmap font size}

\item{anno_colors}{annotation colors}
}
\value{
A ComplexHeatmap heatmap object.
}
\description{
Useful plotting function to plot heatmap of module expression across time with pre-split matrix
}
