% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_heatmap.R
\name{hm_dyn_state}
\alias{hm_dyn_state}
\title{heatmap}
\usage{
hm_dyn_state(
  matrix,
  state_result,
  row_cols,
  limits = c(0, 10),
  toScale = FALSE,
  fontsize_row = 4
)
}
\arguments{
\item{matrix}{expression matrix}

\item{state_result}{result of running findDynGenes}

\item{row_cols}{row_cols}

\item{limits}{limits}

\item{toScale}{toScale}

\item{fontsize_row}{fontsize_row
'}
}
\value{
A ComplexHeatmap heatmap object.
}
\description{
heatmap
}
