% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_adapters.R
\name{map_peak_ranges_to_genome}
\alias{map_peak_ranges_to_genome}
\title{Map peak sequence names to a genome}
\usage{
map_peak_ranges_to_genome(peak_ranges, genome)
}
\arguments{
\item{peak_ranges}{Genomic ranges to map.}

\item{genome}{A genome object with sequence levels and lengths.}
}
\value{
A list of mapped ranges, support flags and original/mapped sequence names.
}
\description{
Map peak sequence names to a genome
}
