% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{metrics}
\alias{metrics}
\alias{metrics,Seurat-method}
\alias{metrics,Network-method}
\title{Metrics}
\usage{
metrics(object, ...)

metrics(object, ...)

\S4method{metrics}{Seurat}(object, network = DefaultNetwork(object), celltypes = NULL, ...)

\S4method{metrics}{Network}(object, celltypes = NULL, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}

\item{network}{network}

\item{celltypes}{cell types to analyze, NULL for all cell types}
}
\description{
Metrics

Get metrics
}
