% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiCSN-package.R
\docType{package}
\name{multiCSN-package}
\alias{multiCSN}
\alias{multiCSN-package}
\title{Inferring Cell-Specific Gene Regulatory Network}
\source{
\url{https://mengxu98.github.io/multiCSN/}
}
\description{
Inferring cell-specific gene regulatory networks from single-cell multiome data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mengxu98.github.io/multiCSN/}
  \item Report bugs at \url{https://github.com/mengxu98/multiCSN/issues}
}

}
\author{
Meng Xu (Maintainer), \email{mengxu98@qq.com}
}
