% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\docType{class}
\name{Network-class}
\alias{Network-class}
\title{The Network class}
\description{
The Network object stores the inferred network itself,
 information about the fitting process as well as graph representations of the network.
}
\section{Slots}{

\describe{
\item{\code{data}}{A matrix.}

\item{\code{modules}}{A list TF modules.}

\item{\code{regulators}}{A named list containing the transcription factors included in the network.}

\item{\code{targets}}{A named list containing the targets included in the network.}

\item{\code{metrics}}{A dataframe with goodness of fit measures.}

\item{\code{coefficients}}{A dataframe with the fitted coefficients.}

\item{\code{graphs}}{Graphical representations of the inferred network.}

\item{\code{network}}{A dataframe containing the network edges and their properties.}

\item{\code{params}}{A named list with GRN inference parameters.}

\item{\code{perturbation}}{A \code{Perturbation} object containing perturbation analysis results.}
}}

