% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{NetworkGraph}
\alias{NetworkGraph}
\alias{NetworkGraph,Seurat-method}
\alias{NetworkGraph,Network-method}
\title{Get network graph}
\usage{
NetworkGraph(object, ...)

NetworkGraph(object, ...)

\S4method{NetworkGraph}{Seurat}(
  object,
  network = DefaultNetwork(object),
  graph = "module_graph",
  celltypes = NULL,
  ...
)

\S4method{NetworkGraph}{Network}(object, graph = "module_graph", ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}

\item{network}{network}

\item{graph}{graph}

\item{celltypes}{cell types to analyze, NULL for all cell types}
}
\description{
Get network graph

Get network graph
}
