% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{NetworkParams}
\alias{NetworkParams}
\alias{NetworkParams,Network-method}
\alias{NetworkParams,Seurat-method}
\title{Get network parameters}
\usage{
NetworkParams(object, ...)

NetworkParams(object, ...)

\S4method{NetworkParams}{Network}(object, ...)

\S4method{NetworkParams}{Seurat}(object, network = DefaultNetwork(object), celltypes = NULL, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}

\item{network}{network}

\item{celltypes}{cell types to analyze, NULL for all cell types}
}
\description{
Get network parameters

Get network parameters
}
