% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{NetworkTFs}
\alias{NetworkTFs}
\alias{NetworkTFs,Seurat-method}
\title{Get network TFs}
\usage{
NetworkTFs(object, ...)

NetworkTFs(object, ...)

\S4method{NetworkTFs}{Seurat}(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}
}
\description{
Get network TFs

Get network TFs
}
