% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{Params}
\alias{Params}
\alias{Params,Seurat-method}
\title{Get parameters}
\usage{
Params(object, ...)

Params(object, ...)

\S4method{Params}{Seurat}(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}
}
\description{
Get parameters

Get GRN inference parameters
}
