% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_heatmap.R
\name{plot_detail_network}
\alias{plot_detail_network}
\title{Plot the dynamic differential network but colored by communities and optionally faded by igraph::betweenness}
\usage{
plot_detail_network(
  network,
  regulators,
  top_edges = NULL,
  only_TFs = TRUE,
  network_order = NULL,
  communities = NULL,
  compute_betweenness = TRUE
)
}
\arguments{
\item{network}{the dynamic network}

\item{regulators}{regulators}

\item{top_edges}{top_edges}

\item{only_TFs}{whether or not to only plot regulators and exclude non-regulators}

\item{network_order}{the network_order in which to plot states, or which states to plot}

\item{communities}{community assignments or the result of running find_commumities.
The names in this object should match the names of the state networks in network.
If NULL, it will be automatically run.}

\item{compute_betweenness}{whether or not to fade nodes by igraph::betweenness}
}
\value{
plot
}
\description{
Plot the dynamic differential network but colored by communities and optionally faded by igraph::betweenness
}
