% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_comparison.R
\name{plot_dyn_diffnet}
\alias{plot_dyn_diffnet}
\title{Plot the dynamic differential network}
\usage{
plot_dyn_diffnet(grn, tfs, only_TFs = TRUE, order = NULL)
}
\arguments{
\item{grn}{the dynamic network}

\item{tfs}{TFs}

\item{only_TFs}{whether or not to only plot TFs and exclude non-regulators}

\item{order}{the order in which to plot states, or which states to plot}
}
\value{
plot
}
\description{
Plot the dynamic differential network
}
