% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_heatmap.R
\name{plot_dynamic_network}
\alias{plot_dynamic_network}
\title{quick plot of dynamic networks}
\usage{
plot_dynamic_network(
  network,
  regulators,
  only_TFs = TRUE,
  network_order = NULL,
  weight_threshold = NULL
)
}
\arguments{
\item{network}{the result of running split_network_by_states}

\item{regulators}{regulators}

\item{only_TFs}{plot only regulator network}

\item{network_order}{which states or transitions to plot}

\item{weight_threshold}{weight_threshold}
}
\value{
plot
}
\description{
quick plot of dynamic networks
}
