% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/generics.R, R/utils_plots.R
\name{plot_gof}
\alias{plot_gof}
\alias{plot_gof,Seurat-method}
\title{Plot goodness of fit}
\usage{
plot_gof(object, ...)

plot_gof(object, ...)

\S4method{plot_gof}{Seurat}(object, network = NULL, celltypes = NULL, point_size = 0.5, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Additional arguments.}

\item{network}{Name of the network to use.}

\item{celltypes}{Celltypes to plot.
If \code{NULL}, all celltypes are plotted.}

\item{point_size}{Float indicating the point size.}
}
\value{
A ggplot2 object.
}
\description{
Plot goodness of fit

Plot goodness-of-fit

Plot goodness-of-fit metrics.
}
