% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_analysis.R
\name{plot_key_tf_multiome}
\alias{plot_key_tf_multiome}
\title{Plot integrated multiome key-TF scores}
\usage{
plot_key_tf_multiome(tf_scores, top_n = 12)
}
\arguments{
\item{tf_scores}{Output of \code{prioritize_features(type = "regulators", strategy = "integrated")}.}

\item{top_n}{Number of TFs shown.}
}
\value{
A patchwork/ggplot object.
}
\description{
Plot integrated multiome key-TF scores
}
