% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/generics.R, R/utils_plots.R
\name{plot_module_metrics}
\alias{plot_module_metrics}
\alias{plot_module_metrics,Seurat-method}
\title{Plot module metrics}
\usage{
plot_module_metrics(object, ...)

plot_module_metrics(object, ...)

\S4method{plot_module_metrics}{Seurat}(object, network = NULL, celltypes = NULL, ...)
}
\arguments{
\item{object}{CSNObject or Network}

\item{...}{Additional arguments}

\item{network}{Name of the network to use.}

\item{celltypes}{Celltypes to plot.
If \code{NULL}, all celltypes are plotted.}
}
\value{
A ggplot2 object.
}
\description{
Plot module metrics

Plot module metrics

Plot module metrics number of genes, number of peaks and number of TFs per gene.
}
