% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_heatmap.R
\name{plot_targets_and_regulators}
\alias{plot_targets_and_regulators}
\title{Updated plot of top regulators given targets in dynamic networks based on a weight column.
Top regulators computed for each state, but maintained in plot across states if present in state subnetwork.}
\usage{
plot_targets_and_regulators(
  network,
  targets,
  states = NULL,
  weight_column = "weight",
  gene_ranks = NULL,
  regulators_num = 5,
  network_order = NULL,
  fixed_layout = TRUE,
  declutter = TRUE,
  show_expression = TRUE,
  node_size = 5,
  label_size = 2,
  title_size = 10,
  legend_size = 8
)
}
\arguments{
\item{network}{the result of running split_network_by_states}

\item{targets}{targets}

\item{states}{result of running assign_genes_to_states}

\item{weight_column}{column name containing reconstruction weights to use}

\item{gene_ranks}{gene_ranks}

\item{regulators_num}{number of top regulators to plot}

\item{network_order}{which states or transitions to plot}

\item{fixed_layout}{whether or not to fix node positions across state networks}

\item{declutter}{if TRUE, will only label nodes with active interactions in given network}

\item{show_expression}{if TRUE, size and shade of node indicates mean expression in a given state.}

\item{node_size}{node_size}

\item{label_size}{label_size}

\item{title_size}{title_size}

\item{legend_size}{legend_size}
}
\value{
plot
}
\description{
Updated plot of top regulators given targets in dynamic networks based on a weight column.
Top regulators computed for each state, but maintained in plot across states if present in state subnetwork.
}
