% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/generics.R, R/utils_plots.R
\name{plot_tf_network}
\alias{plot_tf_network}
\alias{plot_tf_network,Seurat-method}
\title{Plot TF network}
\usage{
plot_tf_network(object, ...)

plot_tf_network(object, ...)

\S4method{plot_tf_network}{Seurat}(
  object,
  tfs = NULL,
  network = NULL,
  celltypes = NULL,
  graph = "module_graph",
  circular = TRUE,
  edge_width = 0.2,
  edge_color = c(`-1` = "darkgrey", `1` = "orange"),
  node_size = 3,
  text_size = 10,
  label_nodes = c("tfs", "all", "none"),
  color_edges = TRUE,
  ...
)
}
\arguments{
\item{object}{CSNObject or Network}

\item{...}{Additional arguments}

\item{tfs}{The transcription factor to center around.}

\item{network}{Name of the network to use.}

\item{celltypes}{Celltypes to plot.
If \code{NULL}, all celltypes are plotted.}

\item{graph}{Name of the graph.}

\item{circular}{Logical. Layout tree in circular layout.}

\item{edge_width}{Edge width.}

\item{edge_color}{Edge color.}

\item{node_size}{Node size.}

\item{text_size}{Font size for labels.}

\item{label_nodes}{String, indicating what to label.
* \code{'tfs'} - Label all TFs.
* \code{'all'} - Label all genes.
* \code{'none'} - Label nothing (except the root TF).}

\item{color_edges}{Logical, whether to color edges by direction.}
}
\value{
A CSNObject object.
}
\description{
Plot TF network

Plot TF network

Plot sub-network centered around one TF.
}
