% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/network_perturbation.R
\name{plotPerturbation}
\alias{plotPerturbation}
\alias{plotPerturbation,Network-method}
\alias{plotPerturbation,Seurat-method}
\alias{plotPerturbation,CSNObject-method}
\title{Plot perturbation}
\usage{
plotPerturbation(
  object,
  label_points = FALSE,
  point_size = 1,
  max_overlaps = 10,
  alpha = 0.6,
  show_trajectory = TRUE,
  arrow_size = 0.5,
  grid_n = 25
)

plotPerturbation(
  object,
  label_points = FALSE,
  point_size = 1,
  max_overlaps = 10,
  alpha = 0.6,
  show_trajectory = TRUE,
  arrow_size = 0.5,
  grid_n = 25
)

\S4method{plotPerturbation}{Network}(
  object,
  label_points = FALSE,
  point_size = 1,
  max_overlaps = 10,
  alpha = 0.6,
  show_trajectory = TRUE,
  arrow_size = 0.5,
  grid_n = 25
)

\S4method{plotPerturbation}{Seurat}(
  object,
  label_points = FALSE,
  point_size = 1,
  max_overlaps = 10,
  alpha = 0.6,
  show_trajectory = TRUE,
  arrow_size = 0.5,
  grid_n = 25
)

\S4method{plotPerturbation}{CSNObject}(
  object,
  label_points = FALSE,
  point_size = 1,
  max_overlaps = 10,
  alpha = 0.6,
  show_trajectory = TRUE,
  arrow_size = 0.5,
  grid_n = 25
)
}
\arguments{
\item{object}{A Network or CSNObject object}

\item{label_points}{Logical, whether to add point labels}

\item{point_size}{Numeric, size of points}

\item{max_overlaps}{Maximum number of overlapping labels to allow}

\item{alpha}{Point transparency}

\item{show_trajectory}{Logical, whether to show trajectory arrows}

\item{arrow_size}{Size of trajectory arrows}

\item{grid_n}{Number of grid points for trajectory arrows}
}
\value{
ggplot object
}
\description{
Plot perturbation

Visualize perturbation results
}
