% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/network_perturbation.R
\name{plotPerturbationTrajectory}
\alias{plotPerturbationTrajectory}
\alias{plotPerturbationTrajectory,Network-method}
\alias{plotPerturbationTrajectory,Seurat-method}
\title{Plot perturbation trajectory}
\usage{
plotPerturbationTrajectory(
  object,
  genes = NULL,
  show_heatmap = FALSE,
  n_top_genes = 50,
  heatmap_style = "combined",
  ...
)

plotPerturbationTrajectory(
  object,
  genes = NULL,
  show_heatmap = FALSE,
  n_top_genes = 50,
  heatmap_style = "combined",
  ...
)

\S4method{plotPerturbationTrajectory}{Network}(
  object,
  genes = NULL,
  show_heatmap = FALSE,
  n_top_genes = 50,
  heatmap_style = "combined",
  ...
)

\S4method{plotPerturbationTrajectory}{Seurat}(
  object,
  genes = NULL,
  show_heatmap = FALSE,
  n_top_genes = 50,
  heatmap_style = "combined",
  ...
)
}
\arguments{
\item{object}{Network or CSNObject}

\item{genes}{Genes to plot}

\item{show_heatmap}{Logical, whether to show expression heatmap}

\item{n_top_genes}{Number of top variable genes to show in heatmap}

\item{heatmap_style}{Style of heatmap visualization ("combined" or "separate")}

\item{...}{Additional arguments}
}
\value{
ggplot object or list of plots
}
\description{
Plot perturbation trajectory

Plot perturbation trajectory
}
