% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiCSN-package.R
\name{print.multicsn_logo}
\alias{print.multicsn_logo}
\title{Print logo}
\usage{
\method{print}{multicsn_logo}(x, ...)
}
\arguments{
\item{x}{Input information.}

\item{...}{Other parameters.}
}
\value{
Print the ASCII logo
}
\description{
Print logo
}
