% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{rough_hierarchy}
\alias{rough_hierarchy}
\title{rough_hierarchy}
\usage{
rough_hierarchy(network_table, abs_weight = TRUE, directed = TRUE)
}
\arguments{
\item{network_table}{The weight data table of network.}

\item{abs_weight}{Whether to use absolute weight values.}

\item{directed}{Whether the network is directed.}
}
\value{
list
}
\description{
returns rough roots in the network, rough roots selected as those connected to most number of nodes
}
\examples{
data("example_matrix", package = "inferCSN")
network_table <- inferCSN::inferCSN(example_matrix)
rough_hierarchy(network_table)
}
