% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_analysis.R
\name{select_key_tfs}
\alias{select_key_tfs}
\title{Select top prioritized TFs}
\usage{
select_key_tfs(
  object,
  network = DefaultNetwork(object),
  pseudotime_column = NULL,
  assay = NULL,
  peak_assay = NULL,
  method = c("page_rank", "degree_distribution"),
  require_all_states = TRUE,
  min_state_fraction = 1,
  min_targets = 10,
  top_n = 6
)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{network}{Dynamic or static network name. Defaults to \code{DefaultNetwork(object)}.}

\item{pseudotime_column}{Optional pseudotime column.}

\item{assay}{RNA assay used for TF expression summaries.}

\item{peak_assay}{Chromatin assay used for motif-accessibility summaries.}

\item{method}{Regulator ranking method.}

\item{require_all_states}{Require TFs to be present in all inferable states.}

\item{min_state_fraction}{Minimum fraction of states where TF must be present
when \code{require_all_states = FALSE}.}

\item{min_targets}{Minimum number of targets in at least one state.}

\item{top_n}{Number of TFs returned.}
}
\value{
Character vector of TFs.
}
\description{
Select top prioritized TFs
}
