% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{ShortestPaths}
\alias{ShortestPaths}
\alias{ShortestPaths,Seurat-method}
\alias{ShortestPaths,Network-method}
\title{Get stored shortest-path results from Network/CSNObject}
\usage{
ShortestPaths(object, ...)

\S4method{ShortestPaths}{Seurat}(object, network = DefaultNetwork(object), celltypes = NULL, ...)

\S4method{ShortestPaths}{Network}(object, ...)
}
\arguments{
\item{object}{The input data.}

\item{...}{Arguments for other methods}

\item{network}{Name of the network.}

\item{celltypes}{Cell types/states to get paths for; NULL for all.}
}
\description{
Get stored shortest-path results from Network/CSNObject
}
