% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_network.R
\name{split_network_by_states}
\alias{split_network_by_states}
\title{Divides grnDF into states, filters interactions between genes not in same or consecutive states}
\usage{
split_network_by_states(grnDF, states, state_network = NULL)
}
\arguments{
\item{grnDF}{result of GRN reconstruction}

\item{states}{result of running assign_genes_to_states}

\item{state_network}{dataframe outlining higher level state connectivity (i.e. state transition network).
If NULL, will assume states is ordered linear trajectory}
}
\value{
list of GRNs across states and transitions
}
\description{
Divides grnDF into states, filters interactions between genes not in same or consecutive states
}
