% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_network.R
\name{split_states_by_group}
\alias{split_states_by_group}
\title{Splits data into states manually}
\usage{
split_states_by_group(dynamic_object, assignment)
}
\arguments{
\item{dynamic_object}{result of running findDynGenes or compileDynGenes}

\item{assignment}{a list of vectors where names(assignment) are state names, and vectors contain groups belonging to corresponding state}
}
\value{
updated dynamic_object with state column included in dynamic_object$cells
}
\description{
Splits data into states given group assignment
}
