% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_network.R
\name{split_states_by_pseudotime}
\alias{split_states_by_pseudotime}
\title{Splits data into states}
\usage{
split_states_by_pseudotime(dynamic_object, cuts, state_names = NULL)
}
\arguments{
\item{dynamic_object}{result of running findDynGenes or compileDynGenes}

\item{cuts}{vector of pseudotime cutoffs}

\item{state_names}{names of resulting states, must have length of length(cuts)+1}
}
\value{
updated dynamic_object with state column included in dynamic_object$cells
}
\description{
Splits data into states by assigning cells to states
}
