% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{export_csn}
\alias{export_csn}
\alias{export_csn,Seurat-method}
\alias{export_csn,Network-method}
\title{Export CSN}
\usage{
export_csn(object, ...)

export_csn(object, ...)

\S4method{export_csn}{Seurat}(
  object,
  active_network = NULL,
  celltypes = NULL,
  weight_cutoff = NULL,
  ...
)

\S4method{export_csn}{Network}(object, weight_cutoff = NULL, ...)
}
\arguments{
\item{object}{A CSNObject object}

\item{...}{Additional arguments}

\item{active_network}{Character string specifying which network to export}

\item{celltypes}{Character vector of cell types to export}

\item{weight_cutoff}{Numeric threshold for filtering edges by absolute weight}
}
\value{
A list of network data frames by cell type
}
\description{
Export network data from a CSN object with optional filtering
}
\examples{
\dontrun{
data(pbmcmultiome_sub, package = "scop")
data(motifs)
data(motif2tf)

object <- Seurat::NormalizeData(pbmcmultiome_sub, assay = "RNA", verbose = FALSE)
object <- Signac::RunTFIDF(object, assay = "peaks", verbose = FALSE)
object <- initiate_object(
  object,
  group.by = "CellType",
  rna_assay = "RNA",
  peak_assay = "peaks",
  verbose = FALSE
)
genome <- getExportedValue("BSgenome.Hsapiens.UCSC.hg38", "BSgenome.Hsapiens.UCSC.hg38")
object <- find_motifs(
  object,
  pfm = motifs,
  motif_tfs = motif2tf,
  genome = genome,
  backend = "motifmatchr",
  verbose = FALSE
)
object <- inferCSN(object, cores = 2, verbose = FALSE)

networks <- export_csn(object)
head(networks[[1]])
}
}
