% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_modules.R
\name{find_modules}
\alias{find_modules}
\alias{find_modules,Network-method}
\alias{find_modules,Seurat-method}
\alias{find_modules,CSNObject-method}
\title{Find TF modules in regulatory network}
\usage{
find_modules(object, ...)

\S4method{find_modules}{Network}(
  object,
  p_thresh = 0.05,
  rsq_thresh = 0.1,
  nvar_thresh = 10,
  min_genes_per_module = 5,
  xgb_method = c("tf", "target"),
  xgb_top = 50,
  verbose = TRUE,
  ...
)

\S4method{find_modules}{Seurat}(
  object,
  network = NULL,
  p_thresh = 0.05,
  rsq_thresh = 0.1,
  nvar_thresh = 10,
  min_genes_per_module = 5,
  verbose = TRUE,
  ...
)

\S4method{find_modules}{CSNObject}(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Arguments for other methods}

\item{p_thresh}{Float indicating the significance threshold on the adjusted p-value.}

\item{rsq_thresh}{Float indicating the \eqn{R^2} threshold on the adjusted p-value.}

\item{nvar_thresh}{Integer indicating the minimum number of variables in the model.}

\item{min_genes_per_module}{Integer indicating the minimum number of genes in a module.}

\item{xgb_method}{Method to get modules from xgb models
\code{tf} - Choose top targets for each TF.
\code{target} - Choose top TFs for each target gene.}

\item{xgb_top}{Interger indicating how many top targets/TFs to return.}

\item{verbose}{Print messages.}

\item{network}{Name of the network to use.}
}
\value{
A Network object.

A CSNObject object
}
\description{
Find TF modules in regulatory network
}
