% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_model}
\alias{fit_model}
\title{Fit model}
\usage{
fit_model(
  formula,
  data,
  method = c("greedy_l0", "glm", "glmnet", "cv.glmnet", "xgb", "susie"),
  family = gaussian,
  alpha = 1,
  max_support_size = NULL,
  ...
)
}
\arguments{
\item{formula}{An object of class \emph{\code{formula}} with a symbolic description
of the model to be fitted.}

\item{data}{A \emph{\code{data.frame}} containing the variables in the model.}

\item{method}{A character string indicating the method to fit the model.
This can take any of the following choices:
\itemize{
\item \emph{\code{greedy_l0}} - Greedy L0 regression
\item \emph{\code{glm}} - Generalized Linear Model using \emph{\code{stats::glm}}
\item \emph{\code{glmnet}}, \emph{\code{cv.glmnet}} - Regularized GLM using \emph{\code{glmnet::glmnet}}
\item \emph{\code{xgb}} - Gradient Boosting using \emph{\code{xgboost}}
\item \emph{\code{susie}} - SuSiE regression using \emph{\code{susieR}} package
}}

\item{family}{A description of the error distribution and link function.
See \emph{\code{stats::family}} for details.}

\item{alpha}{The elasticnet mixing parameter, between 0 and 1.
See \emph{\code{glmnet::glmnet}} for details.}

\item{max_support_size}{Maximum support size for the greedy-L0 (\code{greedy_l0}) backend.}

\item{...}{Additional parameters passed to the underlying model fitting function.}
}
\value{
A list containing two data frames:
\itemize{
\item \emph{\code{metrics}} - Goodness of fit measures
\item \emph{\code{coefficients}} - Fitted coefficients
}
}
\description{
Fits various types of regression models including generalized linear models,
regularized models, Bayesian models, and machine learning approaches.
}
\examples{
data("example_matrix", package = "inferCSN")
df <- as.data.frame(example_matrix)
fit_model(g1 ~ ., data = df, method = "greedy_l0")
fit_model(g1 ~ ., data = df, method = "glm")
fit_model(g1 ~ ., data = df, method = "glmnet")
fit_model(g1 ~ ., data = df, method = "cv.glmnet")
fit_model(g1 ~ ., data = df, method = "xgb")
fit_model(g1 ~ ., data = df, method = "susie")
}
