% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\docType{methods}
\name{fit_models}
\alias{fit_models}
\alias{fit_models,Seurat-method}
\alias{fit_models,CSNObject-method}
\alias{fit_models,matrix-method}
\title{Fit models for gene expression}
\usage{
fit_models(
  object,
  regulators = NULL,
  targets = NULL,
  network_name = paste0(method, "_network"),
  peak_to_gene_method = c("Signac", "GREAT"),
  upstream = 1e+05,
  downstream = 0,
  extend = 1e+06,
  only_tss = FALSE,
  peak_to_gene_domains = NULL,
  gene_cor_threshold = 0.1,
  peak_cor_threshold = 0.1,
  aggregate_rna_col = NULL,
  aggregate_peaks_col = NULL,
  method = c("greedy_l0", "glm", "glmnet", "cv.glmnet", "xgb", "susie"),
  interaction_term = ":",
  adjust_method = "fdr",
  scale = FALSE,
  verbose = TRUE,
  cores = 1,
  celltype = NULL,
  ...
)

\S4method{fit_models}{Seurat}(
  object,
  regulators = NULL,
  targets = NULL,
  network_name = paste0(method, "_network"),
  peak_to_gene_method = c("Signac", "GREAT"),
  upstream = 1e+05,
  downstream = 0,
  extend = 1e+06,
  only_tss = FALSE,
  peak_to_gene_domains = NULL,
  gene_cor_threshold = 0.1,
  peak_cor_threshold = 0.1,
  aggregate_rna_col = NULL,
  aggregate_peaks_col = NULL,
  method = c("greedy_l0", "glm", "glmnet", "cv.glmnet", "xgb", "susie"),
  interaction_term = ":",
  adjust_method = "fdr",
  scale = FALSE,
  verbose = TRUE,
  cores = 1,
  celltype = NULL,
  ...
)

\S4method{fit_models}{CSNObject}(
  object,
  regulators = NULL,
  targets = NULL,
  network_name = paste0(method, "_network"),
  peak_to_gene_method = c("Signac", "GREAT"),
  upstream = 1e+05,
  downstream = 0,
  extend = 1e+06,
  only_tss = FALSE,
  peak_to_gene_domains = NULL,
  gene_cor_threshold = 0.1,
  peak_cor_threshold = 0.1,
  aggregate_rna_col = NULL,
  aggregate_peaks_col = NULL,
  method = c("greedy_l0", "glm", "glmnet", "cv.glmnet", "xgb", "susie"),
  interaction_term = ":",
  adjust_method = "fdr",
  scale = FALSE,
  verbose = TRUE,
  cores = 1,
  celltype = NULL,
  ...
)

\S4method{fit_models}{matrix}(
  object,
  regulators = NULL,
  targets = NULL,
  network_name = paste0(method, "_network"),
  peak_to_gene_method = c("Signac", "GREAT"),
  upstream = 1e+05,
  downstream = 0,
  extend = 1e+06,
  only_tss = FALSE,
  peak_to_gene_domains = NULL,
  gene_cor_threshold = 0.1,
  peak_cor_threshold = 0.1,
  aggregate_rna_col = NULL,
  aggregate_peaks_col = NULL,
  method = c("greedy_l0", "glm", "glmnet", "cv.glmnet", "xgb", "susie"),
  interaction_term = ":",
  adjust_method = "fdr",
  scale = FALSE,
  verbose = TRUE,
  cores = 1,
  celltype = NULL,
  ...
)
}
\arguments{
\item{object}{The input data for model fitting}

\item{regulators}{A character vector of regulators}

\item{targets}{A character vector of targets}

\item{network_name}{Name for the inferred network}

\item{peak_to_gene_method}{Method to link peaks to genes ("Signac" or "GREAT")}

\item{upstream}{Distance upstream of TSS to consider}

\item{downstream}{Distance downstream of TSS to consider}

\item{extend}{Extension distance for GREAT method}

\item{only_tss}{Whether to measure distance from TSS only}

\item{peak_to_gene_domains}{Custom regulatory domains}

\item{gene_cor_threshold}{Correlation threshold for TFs}

\item{peak_cor_threshold}{Correlation threshold for peaks}

\item{aggregate_rna_col}{Column name for RNA aggregation}

\item{aggregate_peaks_col}{Column name for peaks aggregation}

\item{method}{Statistical method for model fitting}

\item{interaction_term}{Type of interaction variable}

\item{adjust_method}{Method for p-value adjustment}

\item{scale}{Whether to scale the data}

\item{verbose}{Whether to show progress messages}

\item{cores}{Number of cores for parallel processing}

\item{celltype}{Current cell type being processed}

\item{...}{Additional arguments passed to methods}
}
\value{
A list containing model coefficients and fit statistics
}
\description{
Fit statistical models to infer gene regulatory relationships
}
