% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_srm2}
\alias{fit_srm2}
\title{Fit a sparse regression model}
\usage{
fit_srm2(
  formula,
  data,
  verbose = TRUE,
  max_support_size = NULL,
  min_improvement = 1e-10
)
}
\arguments{
\item{formula}{An object of class \emph{\code{formula}} with a symbolic description
of the model to be fitted.}

\item{data}{A \emph{\code{data.frame}} containing the variables in the model.}

\item{verbose}{If \code{TRUE}, show warning messages.}

\item{max_support_size, min_improvement}{See \link[inferCSN:fit_greedy_l0]{inferCSN::fit_greedy_l0}.}
}
\value{
A list containing two data frames:
\itemize{
\item \emph{\code{metrics}} - Goodness of fit measures
\item \emph{\code{coefficients}} - Fitted coefficients with sparse structure
}
}
\description{
Fits a sparse regression model using \code{\link[inferCSN:fit_greedy_l0]{inferCSN::fit_greedy_l0()}}.
}
\examples{
data("example_matrix", package = "inferCSN")
df <- as.data.frame(example_matrix)
fit_srm2(g1 ~ ., data = df)
}
