% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/initiate_object.R
\name{initiate_object}
\alias{initiate_object}
\alias{initiate_object,matrix-method}
\alias{initiate_object,Seurat-method}
\alias{initiate_object,CSNObject-method}
\title{Initiate object}
\usage{
initiate_object(object, ...)

initiate_object(object, ...)

\S4method{initiate_object}{matrix}(object, regulators = character(0), targets = character(0), ...)

\S4method{initiate_object}{Seurat}(
  object,
  celltype = NULL,
  regulators = NULL,
  targets = NULL,
  group.by = NULL,
  filter_mode = c("celltype_specific", "variable", "celltype", "unfiltered"),
  filter_by = c("celltype", "aggregate"),
  rna_assay = "RNA",
  rna_min_pct = 0.1,
  rna_logfc_threshold = 0.25,
  rna_test_method = "wilcox",
  n_variable_genes = 2000,
  peak_assay = NULL,
  peak_min_pct = 0.05,
  peak_logfc_threshold = 0.1,
  peak_test_method = "LR",
  n_variable_peaks = "q5",
  regions = NULL,
  regions_extend = 0,
  exclude_exons = TRUE,
  only_pos = TRUE,
  verbose = TRUE,
  p_value = 0.05,
  ...
)

\S4method{initiate_object}{CSNObject}(object, ...)
}
\arguments{
\item{object}{A Seurat object containing gene expression and/or chromatin accessibility data.}

\item{...}{Additional arguments passed to marker detection functions.}

\item{regulators}{Regulators to consider for CSN inference. Can be a character
vector applied to all cell types, or a named list (names must match cell
types) of character vectors per cell type.}

\item{targets}{Targets to consider for CSN inference. Same format rules as
\code{regulators}.}

\item{celltype}{Selected celltype(s) to analyze. If NULL, all cells will be used.}

\item{group.by}{Column name in Seurat meta.data for cell type information.}

\item{filter_mode}{Filter mode for identifying cell-type specific features.}

\item{filter_by}{When filter_mode is "variable", specify "aggregate" or "celltype".}

\item{rna_assay}{Name of the gene expression assay.}

\item{rna_min_pct, rna_logfc_threshold, rna_test_method, n_variable_genes}{RNA params.}

\item{peak_assay}{Name of the chromatin accessibility assay. If NULL, only RNA.}

\item{peak_min_pct, peak_logfc_threshold, peak_test_method, n_variable_peaks}{Peak params.}

\item{regions}{Candidate regions (GRanges or data frame). If NULL, all peaks.}

\item{regions_extend}{Base pairs to extend regions.}

\item{exclude_exons}{Whether to consider exons for binding site inference.}

\item{only_pos}{Only return positive markers.}

\item{verbose}{Print progress messages.}

\item{p_value}{Significance threshold for filtering features.}
}
\value{
CSNObject object.
}
\description{
Initiate object

Initiate the \code{CSNObject} object
}
\examples{
data(pbmcmultiome_sub, package = "scop")

object <- Seurat::NormalizeData(pbmcmultiome_sub, assay = "RNA", verbose = FALSE)
object <- Signac::RunTFIDF(object, assay = "peaks", verbose = FALSE)

object <- initiate_object(
  object,
  group.by = "CellType",
  rna_assay = "RNA",
  peak_assay = "peaks",
  verbose = FALSE
)
object
}
