% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_analysis.R
\name{plot_celltype_specificity_heatmap}
\alias{plot_celltype_specificity_heatmap}
\title{Plot cell type rewiring heatmap}
\usage{
plot_celltype_specificity_heatmap(
  object,
  network = DefaultNetwork(object),
  celltypes = NULL,
  top_tfs = 20,
  metric = c("rewiring_score", "regulon_jaccard"),
  weight_column = "weight"
)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{network}{Static network name. Defaults to \code{DefaultNetwork(object)}.}

\item{celltypes}{Optional subset of cell types.}

\item{top_tfs}{Number of TFs to display.}

\item{metric}{Heatmap value. Supported values include
\code{"rewiring_score"}, \code{"weighted_regulon_jaccard"},
\code{"binary_rewiring_score"}, and \code{"regulon_jaccard"}.}

\item{weight_column}{Weight column used when exporting networks.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot cell type rewiring heatmap
}
