% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_visualization.R
\name{plot_static_networks}
\alias{plot_static_networks}
\title{Plot dynamic networks}
\usage{
plot_static_networks(
  network_table,
  regulators = NULL,
  targets = NULL,
  legend_position = "right"
)
}
\arguments{
\item{network_table}{The weight data table of network.}

\item{regulators}{A character vector of regulators to include.}

\item{targets}{A character vector of targets to include.}

\item{legend_position}{The position of legend.}
}
\value{
A ggplot2 object
}
\description{
Plot dynamic networks
}
\examples{
data(example_matrix, package = "inferCSN")
network_table <- inferCSN::inferCSN(example_matrix)
plot_static_networks(
  network_table,
  regulators = "g1"
)
plot_static_networks(
  network_table,
  targets = "g1"
)
plot_static_networks(
  network_table,
  regulators = "g2",
  targets = "g3"
)
}
