% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_plots.R
\name{plot_tf_peak_overlap_scenic_style}
\alias{plot_tf_peak_overlap_scenic_style}
\title{Plot TF-supported peak overlap across ordered states}
\usage{
plot_tf_peak_overlap_scenic_style(
  object,
  key_target_tbl,
  tfs,
  pseudotime_column,
  peak_assay = NULL,
  top_targets_per_tf = 6
)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{key_target_tbl}{Output of \code{\link{prioritize_features}} with \code{type = "targets"} and \code{strategy = "integrated"}.}

\item{tfs}{TFs to visualize.}

\item{pseudotime_column}{Pseudotime column used for state assignment.}

\item{peak_assay}{Peak assay name.}

\item{top_targets_per_tf}{Number of target genes retained per TF.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot TF-supported peak overlap across ordered states
}
